test(toast): migrate is-open
